refactor: remove -v2 suffix from social-links module naming
Update all imports to use the canonical social-links module name
instead of social-links-v2, consolidating the social links
implementation into a single consistent module structure
