--- src/features/portfolio/types/social-links-v2.ts
@@ -1,7 +1,0 @@
-export type SocialLink = {
-  name: string
-  title: string
-  icon: React.JSX.Element
-  handle: string
-  href: string
-}
