+++ src/features/portfolio/types/social-links.ts
@@ -1,10 +1,7 @@
-/** @deprecated */
 export type SocialLink = {
-  /** Icon image URL (absolute or path under /public) shown beside the title. */
-  icon: string
+  name: string
   title: string
-  /** Optional handle/username or subtitle displayed under the title. */
-  subtitle?: string
-  /** External profile URL opened when the item is clicked. */
+  icon: React.JSX.Element
+  handle: string
   href: string
 }
